[Interval] re-implement inference of intervals for logic functions (R.I.P. Interval_system)

+++ src/plugins/e-acsl/tests/gmp/oracle/gen_functions_rec.c
@@ -45,11 +45,11 @@ logic ℤ f4(ℤ n) =
   n < 100? f4(n + 1): (n < 0x7fffffffffffffffL? 0x7fffffffffffffffL: 6);
 
 */
-long __gen_e_acsl_f4(int __gen_e_acsl_n_arg);
+unsigned long __gen_e_acsl_f4(int __gen_e_acsl_n_arg);
 
-long __gen_e_acsl_f4_2(long __gen_e_acsl_n_arg);
+unsigned long __gen_e_acsl_f4_2(long __gen_e_acsl_n_arg);
 
-long __gen_e_acsl_f4_3(__e_acsl_mpz_struct * __gen_e_acsl_n_arg);
+unsigned long __gen_e_acsl_f4_3(__e_acsl_mpz_struct * __gen_e_acsl_n_arg);
 
 int main(void)
 {
@@ -120,22 +120,22 @@ int main(void)
   }
   /*@ assert f4(9) > 0; */
   {
-    long __gen_e_acsl_f4_tapp;
+    unsigned long __gen_e_acsl_f4_tapp;
     __gen_e_acsl_f4_tapp = __gen_e_acsl_f4(9);
-    __e_acsl_assert(__gen_e_acsl_f4_tapp > 0L,(char *)"Assertion",
+    __e_acsl_assert(__gen_e_acsl_f4_tapp > 0UL,(char *)"Assertion",
                     (char *)"main",(char *)"f4(9) > 0",29);
   }
   __retres = 0;
   return __retres;
 }
 
-long __gen_e_acsl_f4_3(__e_acsl_mpz_struct * __gen_e_acsl_n_arg)
+unsigned long __gen_e_acsl_f4_3(__e_acsl_mpz_struct * __gen_e_acsl_n_arg)
 {
-  long __retres;
+  unsigned long __retres;
   __e_acsl_mpz_t __gen_e_acsl_n_11;
   __e_acsl_mpz_t __gen_e_acsl__23;
   int __gen_e_acsl_lt_2;
-  long __gen_e_acsl_if_11;
+  unsigned long __gen_e_acsl_if_11;
   __gmpz_init_set(__gen_e_acsl_n_11,
                   (__e_acsl_mpz_struct const *)(__gen_e_acsl_n_arg));
   __gmpz_init_set_si(__gen_e_acsl__23,100L);
@@ -144,7 +144,7 @@ long __gen_e_acsl_f4_3(__e_acsl_mpz_struct * __gen_e_acsl_n_arg)
   if (__gen_e_acsl_lt_2 < 0) {
     __e_acsl_mpz_t __gen_e_acsl__24;
     __e_acsl_mpz_t __gen_e_acsl_add_7;
-    long __gen_e_acsl_f4_tapp_4;
+    unsigned long __gen_e_acsl_f4_tapp_4;
     __gmpz_init_set_si(__gen_e_acsl__24,1L);
     __gmpz_init(__gen_e_acsl_add_7);
     __gmpz_add(__gen_e_acsl_add_7,
@@ -164,7 +164,7 @@ long __gen_e_acsl_f4_3(__e_acsl_mpz_struct * __gen_e_acsl_n_arg)
                                    (__e_acsl_mpz_struct const *)(__gen_e_acsl__25));
     if (__gen_e_acsl_lt_3 < 0) __gen_e_acsl_if_10 = 9223372036854775807UL;
     else __gen_e_acsl_if_10 = 6UL;
-    __gen_e_acsl_if_11 = (long)__gen_e_acsl_if_10;
+    __gen_e_acsl_if_11 = __gen_e_acsl_if_10;
     __gmpz_clear(__gen_e_acsl__25);
   }
   __retres = __gen_e_acsl_if_11;
@@ -173,15 +173,15 @@ long __gen_e_acsl_f4_3(__e_acsl_mpz_struct * __gen_e_acsl_n_arg)
   return __retres;
 }
 
-long __gen_e_acsl_f4_2(long __gen_e_acsl_n_arg)
+unsigned long __gen_e_acsl_f4_2(long __gen_e_acsl_n_arg)
 {
-  long __retres;
-  long __gen_e_acsl_if_13;
+  unsigned long __retres;
+  unsigned long __gen_e_acsl_if_13;
   if (__gen_e_acsl_n_arg < 100L) {
     __e_acsl_mpz_t __gen_e_acsl_n_10;
     __e_acsl_mpz_t __gen_e_acsl__22;
     __e_acsl_mpz_t __gen_e_acsl_add_6;
-    long __gen_e_acsl_f4_tapp_3;
+    unsigned long __gen_e_acsl_f4_tapp_3;
     __gmpz_init_set_si(__gen_e_acsl_n_10,__gen_e_acsl_n_arg);
     __gmpz_init_set_si(__gen_e_acsl__22,1L);
     __gmpz_init(__gen_e_acsl_add_6);
@@ -198,18 +198,18 @@ long __gen_e_acsl_f4_2(long __gen_e_acsl_n_arg)
     unsigned long __gen_e_acsl_if_12;
     if (__gen_e_acsl_n_arg < 9223372036854775807L) __gen_e_acsl_if_12 = 9223372036854775807UL;
     else __gen_e_acsl_if_12 = 6UL;
-    __gen_e_acsl_if_13 = (long)__gen_e_acsl_if_12;
+    __gen_e_acsl_if_13 = __gen_e_acsl_if_12;
   }
   __retres = __gen_e_acsl_if_13;
   return __retres;
 }
 
-long __gen_e_acsl_f4(int __gen_e_acsl_n_arg)
+unsigned long __gen_e_acsl_f4(int __gen_e_acsl_n_arg)
 {
-  long __retres;
-  long __gen_e_acsl_if_15;
+  unsigned long __retres;
+  unsigned long __gen_e_acsl_if_15;
   if (__gen_e_acsl_n_arg < 100) {
-    long __gen_e_acsl_f4_tapp_2;
+    unsigned long __gen_e_acsl_f4_tapp_2;
     __gen_e_acsl_f4_tapp_2 = __gen_e_acsl_f4_2(__gen_e_acsl_n_arg + 1L);
     __gen_e_acsl_if_15 = __gen_e_acsl_f4_tapp_2;
   }
@@ -217,7 +217,7 @@ long __gen_e_acsl_f4(int __gen_e_acsl_n_arg)
     unsigned long __gen_e_acsl_if_14;
     if ((long)__gen_e_acsl_n_arg < 9223372036854775807L) __gen_e_acsl_if_14 = 9223372036854775807UL;
     else __gen_e_acsl_if_14 = 6UL;
-    __gen_e_acsl_if_15 = (long)__gen_e_acsl_if_14;
+    __gen_e_acsl_if_15 = __gen_e_acsl_if_14;
   }
   __retres = __gen_e_acsl_if_15;
   return __retres;
